move component files
